+++ components/middleware/Digital/Outputs/DigitalOutputs.h
@@ -78,9 +78,9 @@ public:
      * @brief Set the duty cycle value of PWM for a digital output
      *
      * @param num DOUT to set
-     * @param duty Duty cycle, resolution is 14 bits : [0-16.383]
+     * @param duty Duty cycle in percentage [0-100] 
      **/
-    static void setPWMDutyCycle(DOut_Num_t num, uint32_t duty);
+    static void setPWMDutyCycle(DOut_Num_t num, float duty);
 
     /**
      * @brief Get the current of a digital output
